bot/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
aiogram==3.4.1
python-telegram-bot==21.4
